Bug 560292 - Handle -nowarn:[paths] for generated sources
Set the "ignore optional problems" flag correctly on the
ICompilationUnit created for a generated source.

Fix the parsing of the option: for existing paths, remember both the
absolute and the canonical path. The compilation units use the absolute
path, so if only the canonical path is remembered, comparisons will
not match for absolute path containing links.

This occurs for instance in tests on Mac: the gen-src directory is
under java.io.tmpdir, which is /var. But on a Mac, /var is a link to
/private/var. Later compilation phases use paths beginning with /var,
but the canonical path remembered is /private/var.

Change-Id: Ie0c55f481828eb169ee310b9b0c91c5f2258d02f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>