Make the compiler baseline configuration configurable by properties
Currently the baseline settings are hardcoded (and always override the
platform parent configuration), this extract these into some properties
that can be overridden from the commandline.

This also adjust the Jenkinsfile to use this new way of selectively
disable replace/compare of the baseline in case of building the
compiler-compiler.

See https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/1395