[DOM To Model] Keep proper type root for source files
Affects ClassFileTests and some ReconcilerTests.
But this might be a workaround: there should be no real need for passing
the cloneCachingContents() to ASTParser as it can do caching on its own.
It should be investigated what requires to pass a clone for
ClassFileTests to succeed, and probably fix the ASTParser accordingly
instead.