Improve perf for resolver
When unit has lambda or method reference, we need to resolve the
bindings to properly index the lambda or reference. However, resolving
binding can take way too much time.
We implement the same trick as what is done with the ECJ-Parser-based
indexer: we trim useless content out of the unit, that is the javadoc,
and method bodies when no functional code is involved.

Note that further improvements could probably be identified, such as
trimming closed sibling blocks or removing further statements.

Should improve
https://github.com/eclipse-jdtls/eclipse-jdt-core-incubator/issues/1161